// rent-acar-angular/src/app/models/administrator.model.ts
export interface Administrator {
    id: number;
    email: string;
    sifra: string;
    ime: string;
    prezime: string;
    jmbg: string;
    drzava: string;
    grad: string;
    adresa: string;
    telefon: string;
    pozicija: string;
}

export interface AdministratorD {
    email: string;
    sifra: string;
    ime: string;
    prezime: string;
    jmbg: string;
    drzava: string;
    grad: string;
    adresa: string;
    telefon: string;
    pozicija: string;
}